[FEATURE] 문의하기 기능 개발 (#437)
* [FIX] 250326 QA 반영 #436

- 답변 등록된 문의 수정/삭제 시 오류 NOTICE로
- 에디터 관련 NOTICE 오류 추가

* [FIX] 250326 QA 반영 #436

- USER 주말 이전에 현재주차 문제집 비공개
- ADMIN 주말 이후 스터디 수정 불가능

* [FIX] 250326 QA 반영 #436

- 문제집 자동 생성 요일 변경(금 -> 목)
- 문제집 자동 생성 API 추가(오류 시 수동 실행용)

* [FIX] 250326 QA 반영 #436

- 문의 상세 조회 시 soft delete 댓글 미반영

* [FIX] 250326 QA 반영 #436

- 커뮤니티 게시글 상세 조회 시 soft delete 댓글 미반영

* [FIX] 250326 QA 반영 #436

- 기관 생성 시 이름 중복이면 NOTICE 에러

+++ module-admin/src/main/java/org/example/domain/board/controller/response/DetailBoardResponse.java
@@ -60,7 +60,7 @@ public class DetailBoardResponse {
   private boolean myLikeYn;
 
   @Schema(description = "댓글 수")
-  private int replyCount;
+  private long replyCount;
 
   @Schema(description = "최종 저장 여부")
   private boolean saveYn;
