+++ module-user/src/main/java/org/example/domain/inquiry/controller/response/DetailInquiryResponse.java
@@ -49,7 +49,7 @@ public class DetailInquiryResponse {
   private boolean myInquiryYn;
 
   @Schema(description = "댓글 수")
-  private int replyCount;
+  private long replyCount;
 
   @Schema(description = "공개 여부")
   private boolean publicYn;
